+++ packages/unified-react-native/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.3](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/unified-react-native@1.0.0-beta.2...@amplitude/unified-react-native@1.0.0-beta.3) (2026-09-23)
+
+**Note:** Version bump only for package @amplitude/unified-react-native
+
+
+
+
+
 # [1.0.0-beta.2](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/unified-react-native@1.0.0-beta.1...@amplitude/unified-react-native@1.0.0-beta.2) (2026-09-21)
 
 **Note:** Version bump only for package @amplitude/unified-react-native